requirements.txt
discord.py==2.4.0
motor==3.3.2
pymongo==4.6.1
aiohttp==3.9.1
PyYAML==6.0.1
pillow==10.1.0
fastapi==0.109.0
uvicorn==0.27.0
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6
yt-dlp==2024.3.10
PyNaCl==1.5.0
imageio-ffmpeg==0.5.1
